agents_backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-ai>=0.2.13",
    "pydantic-ai-slim[duckduckgo]>=0.2.13",
    "pyjwt>=2.10.1",
    "ruff>=0.11.12",
]
